feat: optimization
design issue fixes and custom drag layer optimization
