+++ src/app/globals.css
@@ -11,3 +11,4 @@
 a {
   text-decoration: none;
 }
+
